pyproject.toml
[tool.poetry]
name = "ai-research-assistant"
version = "1.0.0"
description = "Production-ready backend for an AI-powered research paper management and analysis platform"
authors = ["Research Team"]
readme = "README.md"
license = "Proprietary"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.11"

# --- Core Framework ---
fastapi = "^0.104.1"
uvicorn = { version = "^0.24.0", extras = ["standard"] }
python-multipart = "^0.0.6"
python-jose = { version = "^3.3.0", extras = ["cryptography"] }
pydantic = { version = "^2.5.2", extras = ["email"] }
pydantic-settings = "^2.1.0"
starlette = "^0.27.0"

# --- Database ---
prisma = "^0.11.0"
asyncpg = "^0.29.0"
psycopg2-binary = "^2.9.9"
pgvector = "^0.2.4"

# --- AI / ML ---
# NOTE: torch, transformers, sentence-transformers are no longer needed
# for inference – we use the HuggingFace Hub InferenceClient.
huggingface-hub = ">=0.25.0"
httpx = "^0.25.2"
scikit-learn = "^1.3.2"
hdbscan = "^0.8.33"
numpy = "^1.26.2"
scipy = "^1.11.4"

# --- PDF Processing ---
PyMuPDF = "^1.23.7"
"pdfminer.six" = "^20221105"

# --- Async Task Queue ---
celery = { version = "^5.3.6", extras = ["redis"] }
redis = "^5.0.1"
flower = "^2.0.1"

# --- Google Cloud ---
google-cloud-storage = "^2.13.0"
google-cloud-aiplatform = "^1.38.1"

# --- Caching & Performance ---
orjson = "^3.9.10"

# --- Security & Middleware ---
slowapi = "^0.1.9"
python-dotenv = "^1.0.0"
bcrypt = "^4.1.2"

# --- Monitoring & Logging ---
structlog = "^23.2.0"
sentry-sdk = { version = "^1.39.1", extras = ["fastapi"] }
prometheus-fastapi-instrumentator = "^6.1.0"

# --- Utilities ---
tenacity = "^8.2.3"
aiofiles = "^23.2.1"
python-dateutil = "^2.8.2"
tiktoken = "^0.5.2"
networkx = "^3.2.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.23.2"
pytest-cov = "^4.1.0"
httpx = "^0.25.2"
factory-boy = "^3.3.1"
Faker = "^22.0.0"
ruff = "^0.1.9"
mypy = "^1.8.0"

[tool.poetry.scripts]
serve = "uvicorn app.main:app --host 0.0.0.0 --port 8000"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"

# ============================================
# Tool configuration
# ============================================

[tool.ruff]
target-version = "py311"
line-length = 100
select = ["E", "F", "W", "I", "N", "UP", "S", "B", "A", "C4", "RUF"]
ignore = ["S101", "S603", "S607"]

[tool.ruff.format]
quote-style = "double"

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = "-v --strict-markers --tb=short -q"
markers = [
    "unit: Unit tests (no external dependencies)",
    "integration: Integration tests (require database)",
    "api: API endpoint tests",
    "slow: Slow tests (AI model loading, etc.)",
]
